Fix method

+++ SRS_Application/Assets/Scripts/AccountManage/DataManage.cs
@@ -149,9 +149,9 @@ public class UserData
     public string date;
     public string time_turn_off_light; // hh:mm
 
-    public float getValue() {
-        float h = float.Parse(time_turn_off_light.Substring(0, 1));
-        float m = float.Parse(time_turn_off_light.Substring(3, 4));
+    public double getValue() {
+        double h = double.Parse(time_turn_off_light.Substring(0, 1));
+        double m = double.Parse(time_turn_off_light.Substring(3, 4));
         return h + m / 60;
     }
 }
